+++ src/command_modules/azure-cli-network/azure/cli/command_modules/network/_validators.py
@@ -298,37 +298,37 @@ def get_virtual_network_validator(has_type_field=False, allow_none=False, allow_
 # COMMAND NAMESPACE VALIDATORS
 
 def process_ag_listener_create_namespace(namespace): # pylint: disable=unused-argument
-    if not is_valid_resource_id(namespace.frontend_ip):
+    if namespace.frontend_ip and not is_valid_resource_id(namespace.frontend_ip):
         namespace.frontend_ip = _generate_ag_subproperty_id(
             namespace, 'frontendIpConfigurations', namespace.frontend_ip)
 
-    if not is_valid_resource_id(namespace.frontend_port):
+    if namespace.frontend_port and not is_valid_resource_id(namespace.frontend_port):
         namespace.frontend_port = _generate_ag_subproperty_id(
             namespace, 'frontendPorts', namespace.frontend_port)
 
-    if not is_valid_resource_id(namespace.ssl_cert):
+    if namespace.ssl_cert and not is_valid_resource_id(namespace.ssl_cert):
         namespace.ssl_cert = _generate_ag_subproperty_id(
             namespace, 'sslCertificates', namespace.ssl_cert)
 
 def process_ag_http_settings_create_namespace(namespace): # pylint: disable=unused-argument
-    if not is_valid_resource_id(namespace.probe):
+    if namespace.probe and not is_valid_resource_id(namespace.probe):
         namespace.probe = _generate_ag_subproperty_id(
             namespace, 'probes', namespace.probe)
 
 def process_ag_rule_create_namespace(namespace): # pylint: disable=unused-argument
-    if not is_valid_resource_id(namespace.address_pool):
+    if namespace.address_pool and not is_valid_resource_id(namespace.address_pool):
         namespace.address_pool = _generate_ag_subproperty_id(
             namespace, 'backendAddressPools', namespace.address_pool)
 
-    if not is_valid_resource_id(namespace.http_listener):
+    if namespace.http_listener and not is_valid_resource_id(namespace.http_listener):
         namespace.http_listener = _generate_ag_subproperty_id(
             namespace, 'httpListeners', namespace.http_listener)
 
-    if not is_valid_resource_id(namespace.http_settings):
+    if namespace.http_settings and not is_valid_resource_id(namespace.http_settings):
         namespace.http_settings = _generate_ag_subproperty_id(
             namespace, 'backendHttpSettingsCollection', namespace.http_settings)
 
-    if not is_valid_resource_id(namespace.url_path_map):
+    if namespace.url_path_map and not is_valid_resource_id(namespace.url_path_map):
         namespace.url_path_map = _generate_ag_subproperty_id(
             namespace, 'urlPathMaps', namespace.url_path_map)
 
@@ -361,6 +361,20 @@ def process_ag_create_namespace(namespace):
     # process folded parameters
     if namespace.subnet or namespace.virtual_network_name:
         get_subnet_validator(has_type_field=True, allow_new=True)(namespace)
+
+    prefix_usage_error = CLIError('Do not specify --subnet-address-prefix or --vnet-address-prefix'
+                                  ' when using an existing subnet.')
+    if namespace.subnet_address_prefix:
+        if '__SET__' in namespace.subnet_address_prefix:
+            if namespace.subnet_type != 'new':
+                raise prefix_usage_error
+            namespace.subnet_address_prefix = namespace.subnet_address_prefix.replace('__SET__', '')
+
+    if namespace.vnet_address_prefix:
+        if '__SET__' in namespace.vnet_address_prefix:
+            if namespace.subnet_type != 'new':
+                raise prefix_usage_error
+            namespace.vnet_address_prefix = namespace.vnet_address_prefix.replace('__SET__', '')
 
     if namespace.public_ip_address:
         get_public_ip_validator(