+++ src/command_modules/azure-cli-storage/azure/cli/command_modules/storage/_command_type.py
@@ -7,13 +7,13 @@ from azure.cli.core.commands import create_command, command_table
 from azure.cli.command_modules.storage._validators import validate_client_parameters
 
 
-def cli_storage_data_plane_command(name, operation, client_factory,
-                                   transform=None, table_transformer=None):
+def cli_storage_data_plane_command(name, operation, client_factory,  # pylint: disable=too-many-arguments
+                                   transform=None, table_transformer=None, exception_handler=None):
     """ Registers an Azure CLI Storage Data Plane command. These commands always include the
     four parameters which can be used to obtain a storage client: account-name, account-key,
     connection-string, and sas-token. """
-    command = create_command(__name__, name, operation, transform, table_transformer, client_factory)  # pylint: disable=line-too-long
-
+    command = create_command(__name__, name, operation, transform, table_transformer,
+                             client_factory, exception_handler=exception_handler)
     # add parameters required to create a storage client
     group_name = 'Storage Account'
     command.add_argument('account_name', '--account-name', required=False, default=None,