+++ src/command_modules/azure-cli-vm/azure/cli/command_modules/vm/_validators.py
@@ -586,15 +586,14 @@ def process_image_create_namespace(namespace):
 
 
 def _figure_out_storage_source(resource_group_name, source):
-    source = source.lower()
     source_blob_uri = None
     source_disk = None
     source_snapshot = None
     if source.lower().endswith('.vhd'):
         source_blob_uri = source
-    elif '/disks/' in source:
+    elif '/disks/' in source.lower():
         source_disk = source
-    elif '/snapshots/' in source:
+    elif '/snapshots/' in source.lower():
         source_snapshot = source
     else:
         compute_client = _compute_client_factory()
@@ -608,5 +607,18 @@ def _figure_out_storage_source(resource_group_name, source):
 
     return (source_blob_uri, source_disk, source_snapshot)
 
+
+def process_disk_encryption_namespace(namespace):
+    namespace.disk_encryption_keyvault = _get_resource_id(namespace.disk_encryption_keyvault,
+                                                          namespace.resource_group_name,
+                                                          'vaults', 'Microsoft.KeyVault')
+
+    if namespace.key_encryption_keyvault:
+        if not namespace.key_encryption_key:
+            raise CLIError("Incorrect usage '--key-encryption-keyvault': "
+                           "'--key-encryption-key' is required")
+        namespace.key_encryption_keyvault = _get_resource_id(namespace.key_encryption_keyvault,
+                                                             namespace.resource_group_name,
+                                                             'vaults', 'Microsoft.KeyVault')
 
 # endregion