+++ src/command_modules/azure-cli-vm/HISTORY.rst
@@ -3,6 +3,23 @@
 Release History
 ===============
 
+0.1.2rc1 (2017-02-17)
++++++++++++++++++++++
+
+* vm/disk: fix bugs in detach (#2138)
+* Show commands return empty string with exit code 0 for 404 responses (#2117)
+* Disk encryption: Enable/Disable/Show (#2113)
+* vm image: do not normalize casing on blob uri (#2126)
+* vm/av-set: remove domain count defaults (#2111)
+* Move acs commands from vm to acs module (#2098)
+* Fix broken name or ids logic in VM/VMSS Create (#2091)
+* VM list: avoid add None mac addresss (#2059)
+* Use same defaults like portal (#2055)
+* VM: command renaming on 'access' related command (#2053)
+* Add --custom-data to VM and VMSS create (#2035)
+* Improve the default logic on the subnets (#2000)
+* Prompts for yes / no use the -y option rather than --force
+
 0.1.1b3 (2017-02-08)
 +++++++++++++++++++++